+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [4.14.0] - 2020-08-02
 ### `jsonschema-generator`
 #### Added
@@ -436,6 +438,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.14.0...HEAD
 [4.14.0]: https://github.com/victools/jsonschema-generator/compare/v4.13.0...v4.14.0
 [4.13.0]: https://github.com/victools/jsonschema-generator/compare/v4.12.2...v4.13.0
 [4.12.2]: https://github.com/victools/jsonschema-generator/compare/v4.12.1...v4.12.2